chore(types): sync type form docs
