+++ packages/taro/types/api/cloud/index.d.ts
@@ -496,6 +496,7 @@ declare module '../../index' {
     callContainer < R = any, P = any >(params: cloud.CallContainerParam<P>): Promise<cloud.CallContainerResult<R>>
   }
 
+  /** @ignore */
   interface Cloud {
       /** 在调用云开发各 API 前，需先调用初始化方法 init 一次（全局只需一次，多次调用时只有第一次生效）
      * @supported weapp
