+++ packages/taro/types/api/open-api/favorites.d.ts
@@ -31,22 +31,6 @@ declare module '../../index' {
     }
   }
 
-  namespace checkIsAddedToMyMiniProgram {
-    interface Option {
-      /** 接口调用结束的回调函数（调用成功、失败都会执行） */
-      complete?: (res: TaroGeneral.CallbackResult) => void
-      /** 接口调用失败的回调函数 */
-      fail?: (res: TaroGeneral.CallbackResult) => void
-      /** 接口调用成功的回调函数 */
-      success?: (res: TaroGeneral.CallbackResult) => void
-    }
-
-    interface SuccessCallbackResult extends TaroGeneral.CallbackResult {
-      /**是否被添加至 「我的小程序」 */
-      added: boolean
-    }
-  }
-
   interface TaroStatic {
     /** 收藏视频
      * @supported weapp
@@ -80,7 +64,8 @@ declare module '../../index' {
      * @see https://developers.weixin.qq.com/miniprogram/dev/api/open-api/favorites/wx.addVideoToFavorites.html
      */
     addVideoToFavorites(option?: addVideoToFavorites.Option): Promise<TaroGeneral.CallbackResult>
-    /** 预约视频号直播
+
+    /** 收藏文件
      * @supported weapp
      * @example
      * ```tsx
@@ -112,11 +97,5 @@ declare module '../../index' {
      * @see https://developers.weixin.qq.com/miniprogram/dev/api/open-api/favorites/wx.addFileToFavorites.html
      */
     addFileToFavorites(option?: addFileToFavorites.Option): Promise<TaroGeneral.CallbackResult>
-    /**
-     * 检查小程序是否被添加至 「我的小程序」
-     * @supported weapp
-     * @see https://developers.weixin.qq.com/miniprogram/dev/api/open-api/my-miniprogram/wx.checkIsAddedToMyMiniProgram.html
-     */
-    checkIsAddedToMyMiniProgram(option?: checkIsAddedToMyMiniProgram.Option):Promise<checkIsAddedToMyMiniProgram.SuccessCallbackResult>
   }
 }
